requirements.txt
discord.py>=2.3.1
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
numpy>=1.22.3
aiohttp>=3.8.5
apscheduler>=3.10.1
torch>=2.0.0
pathlib>=1.0.1
regex>=2023.0.0
uuid>=1.30
google-generativeai>=0.3.0
rank_bm25
python-docx>=1.1.0
markdownify>=0.11.6
beautifulsoup4>=4.9.0
